+++ android/src/main/java/com/jiguang/jpush/JPushEventReceiver.java
@@ -14,18 +14,27 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import cn.jpush.android.api.CmdMessage;
 import cn.jpush.android.api.JPushInterface;
 import cn.jpush.android.api.JPushMessage;
 import cn.jpush.android.api.NotificationMessage;
-import cn.jpush.android.service.JPushMessageService;
+import cn.jpush.android.service.JPushMessageReceiver;
 import io.flutter.plugin.common.MethodChannel.Result;
 
-public class JPushEventReceiver extends JPushMessageService {
-
+public class JPushEventReceiver extends JPushMessageReceiver {
+    @Override
+    public void onCommandResult(Context context,final CmdMessage cmdMessage) {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
+            @Override
+            public void run() {
+                JPushHelper.getInstance().onCommandResult(cmdMessage);
+            }
+        });
+    }
     @Override
     public void onNotifyMessageUnShow(Context context,final NotificationMessage notificationMessage) {
         super.onNotifyMessageUnShow(context,notificationMessage);
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
             @Override
             public void run() {
                 JPushHelper.getInstance().onNotifyMessageUnShow(notificationMessage);
@@ -35,7 +44,7 @@ public class JPushEventReceiver extends JPushMessageService {
     @Override
     public void onConnected(Context context,final boolean isConnected) {
         //连接状态
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
             @Override
             public void run() {
                 JPushHelper.getInstance().onConnected(isConnected);
@@ -46,7 +55,7 @@ public class JPushEventReceiver extends JPushMessageService {
     @Override
     public void onInAppMessageShow(Context context,final NotificationMessage message) {
         Log.i("JPushPlugin", "[onInAppMessageShow], " + message.toString());
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
             @Override
             public void run() {
                 JPushHelper.getInstance().onInAppMessageShow(message);
@@ -57,7 +66,7 @@ public class JPushEventReceiver extends JPushMessageService {
     @Override
     public void onInAppMessageClick(Context context,final NotificationMessage message) {
         Log.i("JPushPlugin", "[onInAppMessageClick], " + message.toString());
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
             @Override
             public void run() {
                 JPushHelper.getInstance().onInAppMessageClick(message);
@@ -84,7 +93,7 @@ public class JPushEventReceiver extends JPushMessageService {
             return;
         }
 
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
             @Override
             public void run() {
                 if (jPushMessage.getErrorCode() == 0) { // success
@@ -126,7 +135,7 @@ public class JPushEventReceiver extends JPushMessageService {
             return;
         }
 
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
             @Override
             public void run() {
                 if (jPushMessage.getErrorCode() == 0) {
@@ -159,7 +168,7 @@ public class JPushEventReceiver extends JPushMessageService {
             return;
         }
 
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
+        JPushHelper.getInstance().getHandler().post(new Runnable() {
             @Override
             public void run() {
                 if (jPushMessage.getErrorCode() == 0) { // success