+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 3.0.6
+1、修复BUG
+## 3.0.5
+1、修复BUG
+## 3.0.4
+1、升级 iOS SDK JPush 5.3.0, 升级 android  JPush 5.4.0
+2、开放 setLinkMergeEnable、setGeofenceEnable、setSmartPushEnable、setCollectControl 接口设置
+3、删除 setLbsEnable 接口设置
+## 3.0.3
+1、开放onCommandResult 接口回调
+2、升级适配gradle 7.5
+## 3.0.2
+开放支持动态设置appkey
+## 3.0.1
+升级：升级 android  JPush 5.2.2
 ## 3.0.0
 + 适配flutter3.x, 
 + 升级：升级 iOS SDK JCore4.6.2， JPush 5.2.4